Iterasi 5 - refactoring code

+++ app/Http/Controllers/EDController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers;
 use App\Models\Dokumen;
 use App\Models\Jurusan;
 use App\Models\Prodi;
+use App\Models\Status;
 use App\Models\Tahap;
 use App\Traits\CountdownTrait;
 use App\Traits\FileTrait;
+use App\Traits\TableTrait;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
@@ -17,17 +19,21 @@ class EDController extends Controller
 {
     use CountdownTrait;
     use FileTrait;
+    use TableTrait;
 
-    public function home()
+    public function home(Request $request)
     {
-        $deadline = $this->EDCountdown();
+        $deadline = $this->Countdown('evaluasi');
         $keterangan = 'Semua data';
         $user = Auth::user();
+        $query = Dokumen::where('kategori', 'evaluasi');
+        $statuses = Status::all()->take(7);
+        $kategori = 'evaluasi';
 
         if ($user->role_id == 2) {
-            $data = Dokumen::where('kategori', 'evaluasi')->withWhereHas('prodi.jurusan', function ($query) use ($user) {
+            $query->withWhereHas('prodi.jurusan', function ($query) use ($user) {
                 $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->with('prodi', 'status')->latest('tahun')->paginate(8);
+            })->with('prodi', 'status')->latest('tahun');
             $jurusans = null;
             $prodis = Prodi::where('jurusan_id', $user->user_access_file[0]->jurusan_id)->get();
             $years = Dokumen::where('kategori', 'evaluasi')->withWhereHas('prodi.jurusan', function ($query) use ($user) {
@@ -40,16 +46,35 @@ class EDController extends Controller
                 return redirect()->route('ed_table', $id_ed);
             } else {
                 $years = ($evaluasi_diri) ? Dokumen::where('prodi_id', $user->user_access_file[0]->prodi_id)->latest('tahun')->distinct()->pluck('tahun')->toArray() : null;
-                [$id_evaluasi, $sheetData, $data] = null;
-                return view('evaluasi_diri.table', compact('deadline', 'id_evaluasi', 'sheetData', 'years', 'data'));
+                [$id_evaluasi, $sheetData, $file] = null;
+                return view('evaluasi_diri.table', compact('deadline', 'id_evaluasi', 'sheetData', 'years', 'file'));
             }
         } else {
-            $data = Dokumen::where('kategori', 'evaluasi')->with('prodi.jurusan', 'prodi', 'status')->latest('tahun')->paginate(8);
             $jurusans = Jurusan::all();
             $prodis = Prodi::all();
             $years = Dokumen::where('kategori', 'evaluasi')->latest('tahun')->distinct()->pluck('tahun')->toArray();
         }
-        return view('evaluasi_diri.home', compact('deadline', 'years', 'prodis', 'data', 'jurusans', 'keterangan'));
+
+        if ($request->status) {
+            $query->where('status_id', $request->status);
+        }
+        if ($request->tahun) {
+            $query->where('tahun', $request->tahun);
+            $keterangan = $request->tahun;
+        }
+        if ($request->jurusan) {
+            $query->withWhereHas('prodi.jurusan', function ($query) use ($request) {
+                $query->where('id', $request->jurusan);
+            });
+            $keterangan = Jurusan::find($request->jurusan)->nama_jurusan;
+        }
+        if ($request->prodi) {
+            $query->where('prodi_id', $request->prodi);
+            $keterangan = Prodi::find($request->prodi)->nama_prodi;
+        }
+
+        $data = $query->with('prodi', 'prodi.jurusan', 'status', 'tahap')->latest('tahun')->paginate(8);
+        return view('evaluasi_diri.home', compact('deadline', 'years', 'prodis', 'data', 'jurusans', 'keterangan', 'statuses', 'kategori'));
     }
 
     public function add_action(Request $request)
@@ -61,6 +86,20 @@ class EDController extends Controller
                     'file.mimes' => 'File yang diunggah harus berupa file XLSX.',
                 ]);
 
+            $spreadsheet = IOFactory::load($request->file('file'));
+            $sheet = $spreadsheet->getSheet(0);
+            $columns = ['Standar', 'Kriteria', 'Nilai capaian', 'Sebutan', 'Bobot', 'Nilai Tertimbang', 'Link Bukti'];
+            $fileColumns = [];
+            for ($column = 'C'; $column <= 'I'; $column++) {
+                $cellValue = $sheet->getCell($column . 2)->getValue();
+                array_push($fileColumns, $cellValue);
+            }
+
+            $missingColumns = array_diff($columns, $fileColumns);
+            if (count($missingColumns) > 0) {
+                return back()->with('error', 'Mohon periksa kembali file yang Anda unggah! Kolom yang diperlukan tidak ditemukan: ' . join(', ', $missingColumns));
+            }
+
             $data = Dokumen::where(['kategori' => 'evaluasi', 'prodi_id' => $request->prodi, 'tahun' => $request->tahun])->first();
             if ($data) {
                 $this->DeleteFile($data->file_data);
@@ -68,8 +107,7 @@ class EDController extends Controller
             } else {
                 $prodi = Prodi::find($request->prodi);
             }
-            $extension = $request->file('file')->extension();
-            $path = $this->UploadFile($request->file('file'), "Evaluasi Diri_" . $prodi->nama_prodi . "_" . $request->tahun . "." . $extension);
+            $path = $this->UploadFile($request->file('file'), "Evaluasi Diri_" . $prodi->nama_prodi . "_" . $request->tahun . ".xlsx");
             $eddata = Dokumen::updateOrCreate(
                 ['prodi_id' => $request->prodi, 'tahun' => $request->tahun, 'kategori' => 'evaluasi'],
                 [
@@ -83,28 +121,33 @@ class EDController extends Controller
                 ->log('Menambahkan data ' . basename($eddata->file_data));
             return redirect()->route('ed_home')->with('success', 'File berhasil ditambahkan');
         }
-        return redirect()->route('ed_home')->with('error', 'File gagal ditambahkan');
+        return back()->with('error', 'File gagal ditambahkan');
     }
 
     public function table($id_evaluasi)
     {
-        $data = Dokumen::find($id_evaluasi);
+        $table = $this->EDTable($id_evaluasi);
+        $file = $table[0];
+        $sheetData = $table[1];
         $user = Auth::user();
+        $kategori = 'evaluasi';
+        $deadline = $this->Countdown('evaluasi');
 
-        if (($user->role_id == 3 && $data->prodi_id != $user->user_access_file[0]->prodi_id)) {
+        if (($user->role_id == 3 && $file->prodi_id != $user->user_access_file[0]->prodi_id)) {
             activity()->log('Prohibited access | Mencoba akses data prodi lain');
             return redirect()->route('login')->withErrors(['login_gagal' => 'Anda tidak memiliki akses!']);
-        } elseif ($user->role_id == 2 && $data->prodi->jurusan->id != $user->user_access_file[0]->jurusan_id) {
+        } elseif ($user->role_id == 2 && $file->prodi->jurusan->id != $user->user_access_file[0]->jurusan_id) {
             activity()->log('Prohibited access | Mencoba akses data prodi lain');
             return redirect()->route('login')->withErrors(['login_gagal' => 'Anda tidak memiliki akses!']);
         }
 
-        $file = IOFactory::load(storage_path('app/public/' . $data->file_data));
-        $maxCell = $file->getSheet(0)->getHighestRowAndColumn();
-        $sheetData = $file->getSheet(0)->rangeToArray('A1:' . $maxCell['column'] . $maxCell['row'] - 1);
-        $years = Dokumen::where(['kategori' => 'evaluasi', 'prodi_id' => $data->prodi_id])->latest('tahun')->distinct()->pluck('tahun')->toArray();
-        $deadline = $this->EDCountdown();
-        return view('evaluasi_diri.table', compact('deadline', 'id_evaluasi', 'sheetData', 'years', 'data'));
+        if ($user->role_id == 3) {
+            $years = Dokumen::where(['kategori' => 'evaluasi', 'prodi_id' => $file->prodi_id])->latest('tahun')->distinct()->pluck('tahun')->toArray();
+        } else {
+            $years = null;
+        }
+
+        return view('evaluasi_diri.table', compact('deadline', 'id_evaluasi', 'sheetData', 'years', 'file', 'kategori'));
     }
 
     public function delete($id_evaluasi)
@@ -116,12 +159,17 @@ class EDController extends Controller
                 ->performedOn($file)
                 ->log('Menghapus data ' . basename($file->file_data));
             $file->delete();
+        } else {
+            return back()->with('error', 'Gagal menghapus file');
         }
         return redirect()->route('ed_home')->with('success', 'File berhasil dihapus');
     }
 
     public function change_action(Request $request)
     {
+        $data = Dokumen::find($request->id_evaluasi);
+        $prodi = Prodi::find($request->prodi);
+
         if ($request->hasFile('file')) {
             $request->validate([
                 'file' => 'required|mimes:xlsx',
@@ -129,102 +177,63 @@ class EDController extends Controller
                     'file.mimes' => 'File yang diunggah harus berupa file XLSX.',
                 ]);
 
-            $data = Dokumen::find($request->id_evaluasi);
+            $spreadsheet = IOFactory::load($request->file('file'));
+            $sheet = $spreadsheet->getSheet(0);
+            $columns = ['Standar', 'Kriteria', 'Nilai capaian', 'Sebutan', 'Bobot', 'Nilai Tertimbang', 'Link Bukti'];
+            $fileColumns = [];
+            for ($column = 'C'; $column <= 'I'; $column++) {
+                $cellValue = $sheet->getCell($column . 2)->getValue();
+                array_push($fileColumns, $cellValue);
+            }
+
+            $missingColumns = array_diff($columns, $fileColumns);
+            if (count($missingColumns) > 0) {
+                return back()->with('error', 'Mohon periksa kembali file yang Anda unggah! Kolom yang diperlukan tidak ditemukan: ' . join(', ', $missingColumns));
+            }
+
             $this->DeleteFile($data->file_data);
-            $extension = $request->file('file')->extension();
-            $prodi = Prodi::find($request->prodi);
-            $path = $this->UploadFile($request->file('file'), "Evaluasi Diri_" . $prodi->nama_prodi . "_" . $request->tahun . "." . $extension);
-            Dokumen::updateOrCreate(
-                ['id' => $request->id_evaluasi],
-                [
-                    'prodi_id' => $request->prodi,
-                    'status_id' => 1,
-                    'kategori' => 'evaluasi',
-                    'file_data' => $path,
-                ]
-            );
-            Tahap::updateOrCreate(['dokumen_id' => $data->id, 'status_id' => 1]);
-            activity()
-                ->performedOn($data)
-                ->log('Mengubah data evaluasi diri dengan id ' . $data->id);
-            return redirect()->route('ed_home')->with('success', 'File berhasil diubah');
+            $path = $this->UploadFile($request->file('file'), "Evaluasi Diri_" . $prodi->nama_prodi . "_" . $request->tahun . ".xlsx");
+        } else {
+            $path = "Files/Evaluasi Diri_" . $prodi->nama_prodi . "_" . $data->tahun . ".xlsx";
+            $this->ChangeFileName($data->file_data, $path);
         }
-        return redirect()->route('ed_home')->with('error', 'File gagal diubah');
+
+        $data->update([
+            'prodi_id' => $request->prodi,
+            'status_id' => 1,
+            'kategori' => 'evaluasi',
+            'file_data' => $path,
+        ]);
+        Tahap::updateOrCreate(['dokumen_id' => $data->id, 'status_id' => 1]);
+
+        activity()
+            ->performedOn($data)
+            ->log('Mengubah data evaluasi diri dengan id ' . $data->id);
+        return redirect()->route('ed_home')->with('success', 'File berhasil diubah');
     }
 
     public function filter_year($year)
     {
-        $deadline = $this->EDCountdown();
-        $jurusans = Jurusan::all();
-        $keterangan = $year;
         $user = Auth::user();
-        if ($user->role_id == 2) {
-            $prodis = Prodi::where('jurusan_id', $user->user_access_file[0]->jurusan_id)->get();
-            $years = Dokumen::where('kategori', 'evaluasi')->withWhereHas('prodi.jurusan', function ($query) use ($user) {
-                $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->latest('tahun')->distinct()->pluck('tahun')->toArray();
-            $data = Dokumen::withWhereHas('prodi.jurusan', function ($query) use ($user) {
-                $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->where(['kategori' => 'evaluasi', 'tahun' => $year])->with('prodi', 'status')->latest('tahun')->paginate(8);
-        } elseif ($user->role_id == 3) {
-            $data = Dokumen::where(['kategori' => 'evaluasi', 'prodi_id' => $user->user_access_file[0]->prodi_id, 'tahun' => $year])->first();
-            return redirect()->route('ed_table', $data->id);
-        } else {
-            $data = Dokumen::where(['kategori' => 'evaluasi', 'tahun' => $year])->with('prodi', 'prodi.jurusan')->latest('tahun')->paginate(8);
-            $prodis = Prodi::all();
-            $years = Dokumen::where('kategori', 'evaluasi')->latest('tahun')->distinct()->pluck('tahun')->toArray();
-        }
-        return view('evaluasi_diri.home', compact('deadline', 'data', 'years', 'prodis', 'jurusans', 'keterangan'));
-    }
-
-    public function filter_prodi($prodi_id)
-    {
-        $jurusans = Jurusan::all();
-        $deadline = $this->EDCountdown();
-        $user = Auth::user();
-        if ($user->role_id == 2) {
-            $prodis = Prodi::where('jurusan_id', $user->user_access_file[0]->jurusan_id)->get();
-            $years = Dokumen::where('kategori', 'evaluasi')->withWhereHas('prodi.jurusan', function ($query) use ($user) {
-                $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->latest('tahun')->distinct()->pluck('tahun')->toArray();
-            $data = Dokumen::withWhereHas('prodi.jurusan', function ($query) use ($user) {
-                $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->where(['kategori' => 'evaluasi', 'prodi_id' => $prodi_id])->with('prodi', 'status')->latest('tahun')->paginate(8);
-        } else {
-            $prodis = Prodi::all();
-            $years = Dokumen::where('kategori', 'evaluasi')->latest('tahun')->distinct()->pluck('tahun')->toArray();
-            $data = Dokumen::where(['kategori' => 'evaluasi', 'prodi_id' => $prodi_id])->with('prodi', 'prodi.jurusan')->latest('tahun')->paginate(8); //dipersingkat (join dihilangkan) karena menggunakan eloquent relationship
-        }
-        $keterangan = ($data->count()) ? $data[0]->prodi->nama_prodi : 'Data kosong';
-        return view('evaluasi_diri.home', compact('deadline', 'data', 'years', 'prodis', 'jurusans', 'keterangan'));
-    }
-
-    public function filter_jurusan($jurusan_id)
-    {
-        $jurusans = Jurusan::all();
-        $deadline = $this->EDCountdown();
-        $prodis = Prodi::all();
-        $years = Dokumen::where('kategori', 'evaluasi')->latest('tahun')->distinct()->pluck('tahun')->toArray();
-        $data = Dokumen::where('kategori', 'evaluasi')->withWhereHas('prodi.jurusan', function ($query) use ($jurusan_id) {
-            $query->where('id', $jurusan_id);
-        })->with('prodi', 'status')->latest('tahun')->paginate(8);
-        $keterangan = ($data->count()) ? $data[0]->prodi->jurusan->nama_jurusan : 'Data kosong';
-        return view('evaluasi_diri.home', compact('deadline', 'data', 'years', 'prodis', 'jurusans', 'keterangan'));
+        $data = Dokumen::where(['kategori' => 'evaluasi', 'prodi_id' => $user->user_access_file[0]->prodi_id, 'tahun' => $year])->first();
+        return redirect()->route('ed_table', $data->id);
     }
 
     public function add()
     {
-        $deadline = $this->EDCountdown();
+        $deadline = $this->Countdown('evaluasi');
+        $kategori = 'evaluasi';
         $prodis = Prodi::where('jurusan_id', Auth::user()->user_access_file[0]->jurusan_id)->get();
-        return view('evaluasi_diri.import_form', compact('deadline', 'prodis'));
+        return view('evaluasi_diri.import_form', compact('deadline', 'prodis', 'kategori'));
     }
 
     public function change($id_evaluasi)
     {
-        $deadline = $this->EDCountdown();
+        $deadline = $this->Countdown('evaluasi');
+        $kategori = 'evaluasi';
         $prodis = Prodi::where('jurusan_id', Auth::user()->user_access_file[0]->jurusan_id)->get();
         $data = Dokumen::find($id_evaluasi);
-        return view('evaluasi_diri.change_form', compact('deadline', 'prodis', 'data'));
+        return view('evaluasi_diri.change_form', compact('deadline', 'prodis', 'data', 'kategori'));
     }
 
     public function export_all(Request $request)
