+++ app/Http/Controllers/KSController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers;
 use App\Models\Jurusan;
 use App\Models\Dokumen;
 use App\Models\Prodi;
+use App\Models\Status;
 use App\Models\Tahap;
 use App\Traits\CountdownTrait;
 use App\Traits\FileTrait;
+use App\Traits\TableTrait;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
@@ -17,17 +19,21 @@ class KSController extends Controller
 {
     use CountdownTrait;
     use FileTrait;
+    use TableTrait;
 
-    public function home()
+    public function home(Request $request)
     {
-        $deadline = $this->KSCountdown();
+        $deadline = $this->Countdown('standar');
         $keterangan = 'Semua data';
         $user = Auth::user();
+        $query = Dokumen::where('kategori', 'standar');
+        $statuses = Status::all()->take(7);
+        $kategori = 'standar';
 
         if ($user->role_id == 2) {
-            $data = Dokumen::where('kategori', 'standar')->withWhereHas('prodi.jurusan', function ($query) use ($user) {
+            $query->withWhereHas('prodi.jurusan', function ($query) use ($user) {
                 $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->with('prodi', 'status')->latest('tahun')->paginate(8);
+            })->with('prodi', 'status')->latest('tahun');
             $jurusans = null;
             $prodis = Prodi::where('jurusan_id', $user->user_access_file[0]->jurusan_id)->get();
             $years = Dokumen::where('kategori', 'standar')->withWhereHas('prodi.jurusan', function ($query) use ($user) {
@@ -40,16 +46,35 @@ class KSController extends Controller
                 return redirect()->route('ks_table', $id_ed);
             } else {
                 $years = ($ketercapaian_standar) ? Dokumen::where(['kategori' => 'standar', 'prodi_id' => $user->user_access_file[0]->prodi_id])->latest('tahun')->distinct()->pluck('tahun')->toArray() : null;
-                [$id_standar, $sheetData, $headers, $sheetName, $data] = null;
-                return view('ketercapaian_standar.table', compact('deadline', 'id_standar', 'sheetData', 'headers', 'sheetName', 'years', 'data'));
+                [$id_standar, $sheetData, $headers, $sheetName, $file] = null;
+                return view('ketercapaian_standar.table', compact('deadline', 'id_standar', 'sheetData', 'headers', 'sheetName', 'years', 'file'));
             }
         } else {
-            $data = Dokumen::where('kategori', 'standar')->with('prodi.jurusan', 'prodi', 'status')->latest('tahun')->paginate(8);
             $jurusans = Jurusan::all();
             $prodis = Prodi::all();
             $years = Dokumen::where('kategori', 'standar')->latest('tahun')->distinct()->pluck('tahun')->toArray();
         }
-        return view('ketercapaian_standar.home', compact('deadline', 'years', 'prodis', 'data', 'jurusans', 'keterangan'));
+
+        if ($request->status) {
+            $query->where('status_id', $request->status);
+        }
+        if ($request->tahun) {
+            $query->where('tahun', $request->tahun);
+            $keterangan = $request->tahun;
+        }
+        if ($request->jurusan) {
+            $query->withWhereHas('prodi.jurusan', function ($query) use ($request) {
+                $query->where('id', $request->jurusan);
+            });
+            $keterangan = Jurusan::find($request->jurusan)->nama_jurusan;
+        }
+        if ($request->prodi) {
+            $query->where('prodi_id', $request->prodi);
+            $keterangan = Prodi::find($request->prodi)->nama_prodi;
+        }
+
+        $data = $query->with('prodi.jurusan', 'prodi', 'status', 'tahap')->latest('tahun')->paginate(8);
+        return view('ketercapaian_standar.home', compact('deadline', 'years', 'prodis', 'data', 'jurusans', 'keterangan', 'statuses', 'kategori'));
     }
 
     public function add_action(Request $request)
@@ -60,6 +85,19 @@ class KSController extends Controller
             ], [
                     'file.mimes' => 'File yang diunggah harus berupa file XLSX.',
                 ]);
+
+            $spreadsheet = IOFactory::load($request->file('file'));
+            $sheet = $spreadsheet->getSheet(0);
+            $columns = ['Standar', 'NO', 'PERNYATAAN ISI STANDAR ', 'INDIKATOR ', '', '', 'Satuan'];
+            $fileColumns = [];
+            for ($column = 'A'; $column <= 'G'; $column++) {
+                $cellValue = $sheet->getCell($column . 1)->getValue();
+                array_push($fileColumns, $cellValue);
+            }
+            $missingColumns = array_diff($columns, $fileColumns);
+            if (count($missingColumns) > 0) {
+                return back()->with('error', 'Mohon periksa kembali file yang Anda unggah! Kolom yang diperlukan tidak ditemukan: ' . join(', ', $missingColumns));
+            }
 
             $data = Dokumen::where(['kategori' => 'standar', 'prodi_id' => $request->prodi, 'tahun' => $request->tahun])->first();
             if ($data) {
@@ -95,119 +133,64 @@ class KSController extends Controller
                 ->performedOn($file)
                 ->log('Menghapus data ' . basename($file->file_data));
             $file->delete();
+        } else {
+            return back()->with('error', 'Gagal menghapus file');
         }
         return redirect()->route('ks_home')->with('success', 'File berhasil dihapus');
     }
 
     public function table($id_standar)
     {
-        $headers = array();
-        $sheetData = array();
-
+        $table = $this->KSTable($id_standar);
+        $file = $table[0];
+        $headers = $table[1];
+        $sheetCount = $table[2];
+        $sheetName = $table[3];
+        $sheetData = $table[4];
         $user = Auth::user();
-        $data = Dokumen::find($id_standar);
+        $years = Dokumen::where(['kategori' => 'standar', 'prodi_id' => $file->prodi_id])->latest('tahun')->distinct()->pluck('tahun')->toArray();
+        $kategori = 'standar';
+        $deadline = $this->Countdown('standar');
 
-        if (($user->role_id == 3 && $data->prodi_id != $user->user_access_file[0]->prodi_id)) {
+        if (($user->role_id == 3 && $file->prodi_id != $user->user_access_file[0]->prodi_id)) {
             activity()->log('Prohibited access | Mencoba akses data prodi lain');
             return redirect()->route('login')->withErrors(['login_gagal' => 'Anda tidak memiliki akses!']);
-        } elseif ($user->role_id == 2 && $data->prodi->jurusan->id != $user->user_access_file[0]->jurusan_id) {
+        } elseif ($user->role_id == 2 && $file->prodi->jurusan->id != $user->user_access_file[0]->jurusan_id) {
             activity()->log('Prohibited access | Mencoba akses data prodi lain');
             return redirect()->route('login')->withErrors(['login_gagal' => 'Anda tidak memiliki akses!']);
         }
-
-        $file = IOFactory::load(storage_path('app/public/' . $data->file_data));
-        $sheetCount = $file->getSheetCount();
-        $sheetName = $file->getSheetNames();
-        for ($i = 0; $i < $sheetCount - 2; $i++) {
-            $sheet = $file->getSheet($i)->toArray(null, true, true, true);
-            $header = array_shift($sheet);
-
-            array_push($sheetData, $sheet);
-            array_push($headers, $header);
-        }
-        $years = Dokumen::where(['kategori' => 'standar', 'prodi_id' => $data->prodi_id])->latest('tahun')->distinct()->pluck('tahun')->toArray();
-        $deadline = $this->KSCountdown();
-        return view('ketercapaian_standar.table', compact('deadline', 'id_standar', 'sheetData', 'headers', 'sheetName', 'years', 'data'));
+        return view('ketercapaian_standar.table', compact('deadline', 'id_standar', 'sheetData', 'headers', 'sheetName', 'years', 'file', 'kategori'));
     }
 
     public function filter_year($year)
     {
-        $deadline = $this->KSCountdown();
-        $jurusans = Jurusan::all();
-        $keterangan = $year;
         $user = Auth::user();
-        if ($user->role_id == 2) {
-            $prodis = Prodi::where('jurusan_id', $user->user_access_file[0]->jurusan_id)->get();
-            $years = Dokumen::where('kategori', 'standar')->withWhereHas('prodi.jurusan', function ($query) use ($user) {
-                $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->latest('tahun')->distinct()->pluck('tahun')->toArray();
-            $data = Dokumen::withWhereHas('prodi.jurusan', function ($query) use ($user) {
-                $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->where(['kategori' => 'standar', 'tahun' => $year])->with('prodi', 'status')->latest('tahun')->paginate(8);
-        } elseif ($user->role_id == 3) {
-            $data = Dokumen::where(['kategori' => 'standar', 'prodi_id' => $user->user_access_file[0]->prodi_id, 'tahun' => $year])->first();
-            return redirect()->route('ks_table', $data->id);
-        } else {
-            $data = Dokumen::where(['kategori' => 'standar', 'tahun' => $year])->with('prodi', 'prodi.jurusan')->latest('tahun')->paginate(8);
-            $prodis = Prodi::all();
-            $years = Dokumen::where('kategori', 'standar')->latest('tahun')->distinct()->pluck('tahun')->toArray();
-
-        }
-        return view('ketercapaian_standar.home', compact('deadline', 'data', 'years', 'prodis', 'jurusans', 'keterangan'));
-    }
-
-    public function filter_prodi($prodi_id)
-    {
-        $jurusans = Jurusan::all();
-        $deadline = $this->KSCountdown();
-        $user = Auth::user();
-        if ($user->role_id == 2) {
-            $prodis = Prodi::where('jurusan_id', $user->user_access_file[0]->jurusan_id)->get();
-            $years = Dokumen::where('kategori', 'standar')->withWhereHas('prodi.jurusan', function ($query) use ($user) {
-                $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->latest('tahun')->distinct()->pluck('tahun')->toArray();
-            $data = Dokumen::withWhereHas('prodi.jurusan', function ($query) use ($user) {
-                $query->where('id', $user->user_access_file[0]->jurusan_id);
-            })->where(['kategori' => 'standar', 'prodi_id' => $prodi_id])->with('prodi', 'status')->latest('tahun')->paginate(8);
-        } else {
-            $prodis = Prodi::all();
-            $years = Dokumen::where('kategori', 'standar')->latest('tahun')->distinct()->pluck('tahun')->toArray();
-            $data = Dokumen::where(['kategori' => 'standar', 'prodi_id' => $prodi_id])->with('prodi', 'prodi.jurusan')->latest('tahun')->paginate(8);
-        }
-        $keterangan = ($data->count()) ? $data[0]->prodi->nama_prodi : 'Data kosong';
-        return view('ketercapaian_standar.home', compact('deadline', 'data', 'years', 'prodis', 'jurusans', 'keterangan'));
-    }
-
-    public function filter_jurusan($jurusan_id)
-    {
-        $jurusans = Jurusan::all();
-        $deadline = $this->KSCountdown();
-        $prodis = Prodi::all();
-        $years = Dokumen::where('kategori', 'standar')->latest('tahun')->distinct()->pluck('tahun')->toArray();
-        $data = Dokumen::where('kategori', 'standar')->withWhereHas('prodi.jurusan', function ($query) use ($jurusan_id) {
-            $query->where('id', $jurusan_id);
-        })->with('prodi', 'status')->latest('tahun')->paginate(8);
-        $keterangan = ($data->count()) ? $data[0]->prodi->jurusan->nama_jurusan : 'Data kosong';
-        return view('ketercapaian_standar.home', compact('deadline', 'data', 'years', 'prodis', 'jurusans', 'keterangan'));
+        $data = Dokumen::where(['kategori' => 'standar', 'prodi_id' => $user->user_access_file[0]->prodi_id, 'tahun' => $year])->first();
+        return redirect()->route('ks_table', $data->id);
     }
 
     public function add()
     {
-        $deadline = $this->KSCountdown();
+        $deadline = $this->Countdown('standar');
+        $kategori = 'standar';
         $prodis = Prodi::where('jurusan_id', Auth::user()->user_access_file[0]->jurusan_id)->get();
-        return view('ketercapaian_standar.import_form', compact('deadline', 'prodis'));
+        return view('ketercapaian_standar.import_form', compact('deadline', 'prodis', 'kategori'));
     }
 
     public function change($id_standar)
     {
-        $deadline = $this->KSCountdown();
+        $deadline = $this->Countdown('standar');
+        $kategori = 'standar';
         $prodis = Prodi::where('jurusan_id', Auth::user()->user_access_file[0]->jurusan_id)->get();
         $data = Dokumen::find($id_standar);
-        return view('ketercapaian_standar.change_form', compact('deadline', 'prodis', 'data'));
+        return view('ketercapaian_standar.change_form', compact('deadline', 'prodis', 'data', 'kategori'));
     }
 
     public function change_action(Request $request)
     {
+        $data = Dokumen::find($request->id_standar);
+        $prodi = Prodi::find($request->prodi);
+
         if ($request->hasFile('file')) {
             $request->validate([
                 'file' => 'required|mimes:xlsx',
@@ -215,27 +198,38 @@ class KSController extends Controller
                     'file.mimes' => 'File yang diunggah harus berupa file XLSX.',
                 ]);
 
-            $data = Dokumen::find($request->id_standar);
+            $spreadsheet = IOFactory::load($request->file('file'));
+            $sheet = $spreadsheet->getSheet(0);
+            $columns = ['Standar', 'NO', 'PERNYATAAN ISI STANDAR ', 'INDIKATOR ', '', '', 'Satuan'];
+            $fileColumns = [];
+            for ($column = 'A'; $column <= 'G'; $column++) {
+                $cellValue = $sheet->getCell($column . 1)->getValue();
+                array_push($fileColumns, $cellValue);
+            }
+            $missingColumns = array_diff($columns, $fileColumns);
+            if (count($missingColumns) > 0) {
+                return back()->with('error', 'Mohon periksa kembali file yang Anda unggah! Kolom yang diperlukan tidak ditemukan: ' . join(', ', $missingColumns));
+            }
+
             $this->DeleteFile($data->file_data);
-            $extension = $request->file('file')->extension();
-            $prodi = Prodi::find($request->prodi);
-            $path = $this->UploadFile($request->file('file'), "Ketercapaian Standar_" . $prodi->nama_prodi . "_" . $request->tahun . "." . $extension);
-            Dokumen::updateOrCreate(
-                ['id' => $request->id_standar],
-                [
-                    'prodi_id' => $request->prodi,
-                    'status_id' => 1,
-                    'kategori' => 'standar',
-                    'file_data' => $path,
-                ]
-            );
-            Tahap::updateOrCreate(['dokumen_id' => $data->id, 'status_id' => 1]);
-            activity()
-                ->performedOn($data)
-                ->log('Mengubah data ketercapaian standar dengan id ' . $data->id);
-            return redirect()->route('ks_home')->with('success', 'File berhasil diubah');
+            $path = $this->UploadFile($request->file('file'), "Ketercapaian Standar_" . $prodi->nama_prodi . "_" . $request->tahun . ".xlsx");
+        } else {
+            $path = "Files/Ketercapaian Standar_" . $prodi->nama_prodi . "_" . $data->tahun . ".xlsx";
+            $this->ChangeFileName($data->file_data, $path);
         }
-        return redirect()->route('ks_home')->with('error', 'File gagal diubah');
+
+        $data->update([
+            'prodi_id' => $request->prodi,
+            'status_id' => 1,
+            'kategori' => 'standar',
+            'file_data' => $path,
+        ]);
+        Tahap::updateOrCreate(['dokumen_id' => $data->id, 'status_id' => 1]);
+
+        activity()
+            ->performedOn($data)
+            ->log('Mengubah data ketercapaian standar dengan id ' . $data->id);
+        return redirect()->route('ks_home')->with('success', 'File berhasil diubah');
     }
 
     public function export_all(Request $request)
